Update spring-context dependency scope
This change updates the `spring-context` dependency scope to provided.
This is done, so that the dependency is not transitively imported to
projects that use `managed-utils`.

The goal is to have users use their own version of the Spring Framework,
as it suits them. The solution is based on Spring, so the dependency is
expected to be provided at all times.